Adding encryption in sharing files
The files which are sent is being encrypted with base64 before the file is being sent.
On the receiver's end, the file is being decrypted in the receiver computer.
Much needed security update, and planning to make it a bit safer app to release in the internet with more stronger modules.